pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "artha"
version = "5.1.0"
description = "Personal Intelligence OS — AI-powered life management"
readme = "README.md"
license = {text = "AGPL-3.0-or-later"}
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0.1",
    "keyring>=25.0.0",
    "jsonschema>=4.20.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
google = [
    "google-api-python-client>=2.100.0",
    "google-auth>=2.23.0",
    "google-auth-oauthlib>=1.1.0",
    "google-auth-httplib2>=0.1.1",
]
microsoft = [
    "msal>=1.28.0",
    "requests>=2.31.0",
]
apple = [
    "caldav>=1.3.0",
    "vobject>=0.9.6",
]
skills = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
]
channels = [
    # Optional: use python-telegram-bot>=21 for richer Telegram features.
    # The built-in urllib-based adapter (scripts/channels/telegram.py) requires
    # no additional dependencies.
    # "python-telegram-bot>=21.0",
]
all = ["artha[google,microsoft,apple,skills,channels]"]
dev = [
    "pytest>=7.4.0",
    "pytest-mock>=3.11.0",
    "pytest-snapshot>=0.9.0",
    "datadiff>=2.2.0",
    "anyio[trio]>=4.0.0",
]

[project.scripts]
artha = "artha:main"

[project.urls]
Repository = "https://github.com/vedprakash-m/artha"
Documentation = "https://github.com/vedprakash-m/artha/tree/main/docs"

[tool.setuptools]
py-modules = ["artha"]
packages = ["scripts", "scripts.lib", "scripts.skills", "scripts.connectors", "scripts.actions", "scripts.channels", "scripts.middleware", "scripts.schemas"]

# ---------------------------------------------------------------------------
# Ruff — lightweight linter catching bugs that py_compile misses
# ---------------------------------------------------------------------------
[tool.ruff]
target-version = "py311"
line-length = 120
src = ["scripts", "tests"]

[tool.ruff.lint]
select = [
    "F811",   # redefinition of unused name (caught Bug 2: duplicate health_check)
]
# Future: add "F401" (unused import) and "F841" (unused variable) once
# pre-existing violations are cleaned up.
